requirements.txt
pyairtable==2.3.3
python-dotenv==1.0.1
openai==1.54.3
requests==2.32.3
langsmith==0.1.137

